refactor(decision-memory): split replay case building out of replay.py (#242)
The 500-code-line limit puts replay.py at 532. The masking/case side
(load_records through build_cases) moves to replay_cases.py; replay
re-exports it, so preferences_guard and the tests keep their imports.
The dependency runs replay -> replay_cases only.
